Add files via upload
Mini Project

In this project, you are tasked to create a program that automates searching for a word definition in a dictionary. Follow the instructions to implement

Dictionary Project is provided here: https://github.com/mutemip/dictionary-data

We should have a data source (Download from the link above)
Learn how to load json data into a python dictionary
Create a function that returns a definition of a word
Consider a condition that the entered word is not in a dictionary
Consider input from user having different cases – upper/ lower case or mixed eg: RAIN/rain/RaIN
Make your dictionary program more intelligent incase users input a word with wrong spelling the program should be able to suggest the word that might be intended.
eg . pott instead of pot or rainn instead of rain. Tip: use difflib library here